Config and setup management
Cleaned up config directory to be empty at start.
setup script will populate the config directory with appropriate
enabled/disabled versions of the configuration files.

Added setup directory with component parts to be run
when setting up a turnkey system. This includes pre-loading
various software that will be available to choose from
when signing into the network for the first time.